requirements.txt
Django==1.8
Pillow
MySQL-python
django-crispy-forms==1.4.0
python-dateutil
